% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_example-function.R
\name{input_example}
\alias{input_example}
\title{Make an input example for alignontology package}
\usage{
input_example()
}
\value{
output A list of inputs: 1) computed; and 2) reference. Both are
data frames with rows for ontologies and three columns for source, target,
and relation. Feature (source)-ontology (target) relation is annotated as
'feature', while ontology-ontology relation is annotated as 'is_a'. To
differentiate between feature and ontology names, a prefix followed by ':'
precedes an ontology name. The prefix should be different for computed and
reference ontology. All columns are characters.
}
\description{
This function create an input example for alignontology package.
}
\examples{

## Create input example
input=input_example()
}
\keyword{data}
\keyword{example}
